requirements.txt
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.26.0
pyarrow>=15.0.0
requests>=2.31.0
python-dotenv>=1.0.0
ta>=0.11.0
pytz>=2024.1
colorama>=0.4.6
tabulate>=0.9.0
scikit-learn>=1.3.0
alpaca-py>=0.20.0
google-genai>=0.1.0
tavily-python>=0.3.1

# Orchestrator & Dashboard
fastapi
uvicorn
jinja2
python-multipart
psutil
websockets
docker
